refactor(top10-6): split APIModels.swift into 16 focused files under Models/API/
Mechanical MARK-boundary split — no behavioral changes. Each section moves byte-identical into its own file with a 4-line project header and Foundation import. APIModels.swift deleted via git rm. Build clean; 586/586 tests pass (15 skipped, matching baseline).
